[RTL] redirect prints to nanoprint
not sure xbox and nanoprint define `va_list` the same way, this might crash. Also not sure what debugchannel.xbe is, that was just mentioned in CXBX. Maybe the xbox rtl defintion of va_list had something to do with how cxbx treated the va-list version the same as the va-arg one but I doubt it.

To reduce on code size most code that neeeds printf functionality now accesses it through rtl.