Try to implement IoSynchronousDeviceIoControlRequest
just going off the name and a comment in cxbx-r source code and it being right before IoSynchronousFsdRequest in the export table. Seems like maybe some xbox kernel developer noticed that the IRP often didn't need to be adjusted so they added these as helper functions to cut down on boilerplate code?